Support for ESLint v8

+++ tests/dep-time-travel.sh
@@ -13,11 +13,13 @@ npm install --no-save "eslint@${ESLINT_VERSION}" --ignore-scripts
 if [[ -n "$TS_PARSER" ]]; then # if TS parser is manually set, always use it
   echo "Downgrading @typescript-eslint/parser..."
   npm i --no-save "@typescript-eslint/parser@${TS_PARSER}"
-elif [[ "$ESLINT_VERSION" -lt "5" ]]; then # completely remove the new TypeScript parser for ESLint < v5
-  echo "Removing @typescript-eslint/parser..."
-  npm uninstall --no-save @typescript-eslint/parser
-elif [[ "$TRAVIS_NODE_VERSION" -lt "10" ]]; then # TS parser 3 requires node 10+
-  npm i --no-save "@typescript-eslint/parser@3"
+elif [[ "$ESLINT_VERSION" =~ "^[0-9]+$" ]] ; then # workaround for eslint beta.
+  if [[ "$ESLINT_VERSION" -lt "5" ]]; then # completely remove the new TypeScript parser for ESLint < v5
+    echo "Removing @typescript-eslint/parser..."
+    npm uninstall --no-save @typescript-eslint/parser
+  elif [[ "$TRAVIS_NODE_VERSION" -lt "10" ]]; then # TS parser 3 requires node 10+
+    npm i --no-save "@typescript-eslint/parser@3"
+  fi
 fi
 
 # use these alternate TypeScript dependencies for ESLint < v4
